LayoutTests/platform/mac/accessibility/grid-row-authoring-error-ignored-expected.txt

Tests that tables with unignored objects between rows and cells behave as expected.

Grid 1:
AXRole: AXTable AXValue:
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D1
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D1
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXGroup AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C


Grid 2:
AXRole: AXTable AXValue:
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXTable AXValue:
                AXRole: AXRow AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
                AXRole: AXRow AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A1
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A2
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A3
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A1
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A2
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A3
                AXRole: AXGroup AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXTable AXValue:
                AXRole: AXRow AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
                AXRole: AXRow AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A1
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A2
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A3
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A1
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A2
                AXRole: AXColumn AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: A3
                AXRole: AXGroup AXValue:
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: X
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Y
                    AXRole: AXCell AXValue:
                        AXRole: AXStaticText AXValue: Z
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXGroup AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C


Grid 3:
AXRole: AXTable AXValue:
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
    AXRole: AXRow AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D1
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D1
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D2
    AXRole: AXColumn AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: D3
    AXRole: AXGroup AXValue:
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: A
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: B
        AXRole: AXCell AXValue:
            AXRole: AXStaticText AXValue: C


Testing draggable row ignored status:
PASS: !axRow === true
Moving draggable row outside table...
Draggable row no longer ignored.

PASS successfullyParsed is true

TEST COMPLETE
A
B
C
D1
D2
D3
A
B
C
X
Y
Z
A1
A2
A3
D2
D3
A
B
C
D1
D2
D3
